order refs + create temporary file eigenring.py
